Wall running camera tilts based on vertical velocity
Camera tilt varies depending on state of the wall run the character is in. If they just began the wall run, tilt over a fixed time. Once that is over, the camera returns to horizontal based on the player's negative velocity. The camera will also attempt to reset whenever the wall run ends, such as via jumping.